% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_wrap1.R
\name{chemsolve}
\alias{chemsolve}
\title{Mass balance and charge solver}
\usage{
chemsolve(Tc = 300, Nat = 0.2, Kt = 0.2, Clt = 0.4, SO4t = 0.2,
  Cat = 0.1, Mgt = 0.1, start = c(1e-05, 1e-05, 0.15, 0.15, 0.15,
  0.104756881, 0.05, 0.05), maxitr = 100, exprod = NULL, exconstit = NULL,
  exnumz = NULL, excharges = NULL, exa = NULL, exK = NULL)
}
\arguments{
\item{Tc}{Temperature (degrees C)}

\item{Nat}{Sodium concentration (mol/kg)}

\item{Kt}{Potassium (mol/kg)}

\item{Clt}{Chlorine (mol/kg)}

\item{SO4t}{Sulphate (mol/kg)}

\item{Cat}{Calcium (mol/kg)}

\item{Mgt}{Magnesium (mol/kg)}

\item{start}{Initial guess for the calculated equalibrium concentration of the basis species}

\item{maxitr}{Maximum number of iteration}

\item{exprod}{A vector of the names of the additional product species which react to form the basis species}

\item{exconstit}{A vector of the chemical symbols of the additional product species in terms of the basis species}

\item{exnumz}{A vector of the number of basis consituents for each of the additional product species given by the equilibrium equation}

\item{excharges}{A vector of the total charage of the additional product species}

\item{exa}{A vector of the ion size paramters for the additional species}

\item{exK}{A vector of the log K values of the additional product species}
}
\value{
A list containing the concentrations, gamma values, and pH at equilibrium
}
\description{
Mass balance and charge balance solver for chemical equilibria.
}
\details{
A wrapper for the chemsolve_generic function that allow easy addition of product species. If you want to add additional reactant species (i.e. basis species) then the chemsolve_generic function must be used
}
\examples{
## Add H2SO4 as an extra product species

chemsolve(exprod="H2SO4",exconstit="H","H","SO4",exnumz=3,excharges=0,exa=0,exa=-6)

}
