% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_dev.R
\name{htes}
\alias{htes}
\title{EqSolv}
\usage{
htes(Tc = 300, Nat = 0.2, Kt = 0.2, Clt = 0.4, SO4t = 0.2,
  Cat = 0.1, Mgt = 0.1, start = c(1e-05, 1e-05, 0.15, 0.15, 0.15,
  0.104756881, 0.05, 0.05), maxitr = 100)
}
\arguments{
\item{Tc}{Temperature (degrees C)}

\item{Nat}{Sodium}

\item{Kt}{Potassium}

\item{Clt}{Chlorine}

\item{SO4t}{Sulphate}

\item{start}{Initial guess}

\item{maxitr}{Maximum number of iterations}
}
\value{
A list containing the concentrations, gamma values, and pH at equilibrium
}
\description{
Find the equilibrium concentrations
}
