% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_dev2.R
\name{htes2}
\alias{htes2}
\title{EqSolv}
\usage{
htes2(solvent = c("H", "OH"), solvcharge = c("1", "-1"), solva = c("9",
  "4"), Ksoln = -10.908, species = c("Na", "K", "Cl", "SO4", "Ca", "Mg"),
  conc = c(0.2, 0.2, 0.4, 0.2, 0.1, 0.1), a = c(4, 3, 3.5, 4, 6, 8),
  charges = c(1, 1, -1, -2, 2, 2), prod, Tc = 300, start = c(1e-05, 1e-05,
  0.15, 0.15, 0.15, 0.104756881, 0.05, 0.05), maxitr = 100)
}
\arguments{
\item{Tc}{Temperature (degrees C)}

\item{start}{Initial guess}

\item{maxitr}{Maximum number of iterations}

\item{Nat}{Sodium}

\item{Kt}{Potassium}

\item{Clt}{Chlorine}

\item{SO4t}{Sulphate}
}
\value{
A list containing the concentrations, gamma values, and pH at equilibrium
}
\description{
Find the equilibrium concentrations
}
