% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNALL3.R
\name{htes3}
\alias{htes3}
\title{export}
\usage{
htes3(Tc = 300, Nat = 0.2, Kt = 0.2, Clt = 0.4, SO4t = 0.2,
  Cat = 0.1, Mgt = 0.1, start = c(1e-05, 1e-05, 0.15, 0.15, 0.15,
  0.104756881, 0.05, 0.05), maxitr = 100, exprod = NULL, exconstit = NULL,
  exnumz = NULL, excharges = NULL, exa = NULL, exK = NULL)
}
\description{
export
}
