% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemsolve.R
\name{chemsolve}
\alias{chemsolve}
\title{Mass balance and charge solver}
\usage{
chemsolve(Tc = 300, Nat = 0.2, Kt = 0.2, Clt = 0.4, SO4t = 0.2,
  Cat = 0.1, Mgt = 0.1, start = c(1e-05, 1e-05, 0.15, 0.15, 0.15,
  0.104756881, 0.05, 0.05), maxitr = 100, exprod = NULL, exconstit = NULL,
  exnumz = NULL, excharges = NULL, exa = NULL, exK = NULL,
  Clbal = TRUE)
}
\arguments{
\item{Tc}{Temperature (degrees C - between 300 and 400)}

\item{Nat}{Sodium concentration (mol/kg); total}

\item{Kt}{Potassium (mol/kg); total}

\item{Clt}{Chloride (mol/kg); total}

\item{SO4t}{Sulphate (mol/kg); total}

\item{Cat}{Calcium (mol/kg); total}

\item{Mgt}{Magnesium (mol/kg); total}

\item{start}{Initial guess for the calculated equalibrium concentration of the basis species}

\item{maxitr}{Maximum number of iteration
The basis species are:  Na+, K+, Mg2+, Ca2+, Cl-, SO42-. The default complexes are:  NaCl°, KCl°, HCl°, KOH°, NaOH°, KSO4-, NaSO4-,HSO4-,CaSO4°,MgSO4°, MgCl+,CaCl+,CaCl2°,MgOH+,CaOH+.Additional complexes based on the existing basis species are easily added.
If additional complexes are required then these may be added as follows (see example below):}

\item{exprod}{A vector of the names of the complexe(s)}

\item{exconstit}{A vector of the chemical symbol names of the the basis species that are constitute each of the the additional complexes}

\item{exnumz}{A vector of the stiochiometery given by the equilibrium reaction for each of the complexes}

\item{excharges}{A vector of the charge of the complex species}

\item{exa}{A vector of the ion size paramters for the complexes}

\item{exK}{A vector of the log K of the dissociation constants}
}
\value{
A list containing the concentrations, activity coefficients, and pH at equilibrium
}
\description{
Mass balance and charge balance solver for chemical equilibria.
}
\details{
A wrapper for the chemsolve_generic function that allow easy addition of product species.
Use the generic function (chemsolve_generic) if new basis species need to be added or if the log K/temperature range is extended (up or down). \cr
Charge balance is fixed on H+. \cr
Normally total initial moles anions = total moles cations  but excess anions will be balanced by more H+ and vice versa. It is important to choose good initial starting values; for H+, OH- and equilibrium concentrations of the basis species. \cr
Complex dissociation constants (Log K)  are from SupCrt 92  slop98.dat \url{http://geopig.asu.edu/?q=tools} \cr
The Debye_Hückel parameters (A, B & Bdot) equations are polynomial fits to data from tables in Helgeson (1969)  Helgeson & Kirkham (1974) by Nellie Olsen (Note Bdot not used at temperatures  greater than 300°C). \cr
Helgeson H. C. (1969) Thermodynamics of hydrothermal systems at elevated temperatures and pressures. American Journal of Science 267, 729-804. \cr
Helgeson H. C. and Kirkham D. H. (1974) Theoretical prediction of the thermodynamic behavior of aqueous electrolytes at high pressures and temperatures: II. Debye-Hückel parameters for activity coefficients and relative partial molar properties American Journal of Science 274, 1199-1261.
}
\examples{
## Add H2SO4 as an additional complex given the existing list of basis species
chemsolve(exprod="H2SO4",exconstit=c("H","H","SO4"),exnumz=3,excharges=0,exa=0,exK=-6)

##Determine the equilibria at a range of temperatures 
temps <- seq(300,400,10) #A vector of temperatures repeating every 10 degrees from 300 to 400
r <- lapply(temps,chemsolve,Nat=0.2,Kt=0.2,Clt=0.4,SO4t=0.2,Cat=0.1,Mgt=0.1) #Creates a list of the results
r[[1]] #Display results from first temperature
r[[10]] #Display the results of the 10th temperature
}
