% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemsolve.R
\name{chemsolve}
\alias{chemsolve}
\title{Mass balance and charge solver}
\usage{
chemsolve(Tc = 400, Nat = 0.4, Kt = 0.2, Clt = 0.6, SO4t = 0.2,
  Cat = 0.1, Mgt = 0.1, start = c(1e-06, 1e-05, 0.3, 0.1, 0.3, 0.01,
  0.001, 0.02), maxitr = 100, exprod = NULL, exconstit = NULL,
  exnumz = NULL, excharges = NULL, exa = NULL, exK = NULL, bal = "Cl")
}
\arguments{
\item{Tc}{Temperature (degrees C - between 300 and 400)}

\item{Nat}{Sodium concentration (mol/kg); total}

\item{Kt}{Potassium (mol/kg); total}

\item{Clt}{Chloride (mol/kg); total}

\item{SO4t}{Sulphate (mol/kg); total}

\item{Cat}{Calcium (mol/kg); total}

\item{Mgt}{Magnesium (mol/kg); total}

\item{start}{Initial guess for the calculated equilbrium concentrations of the basis species (in the order of H+, OH-, Na+, K+, Cl-, SO42+, Ca2+, Mg2+).}

\item{maxitr}{Maximum number of iterations}

\item{exprod}{A vector of the names of the additional complex(es)}

\item{exconstit}{A vector of the chemical symbol names of the the basis species that are constitute each of the the additional complexes}

\item{exnumz}{A vector of the stiochiometry given by the equilibrium reaction for each of the additional complexes}

\item{excharges}{A vector of the charge of the additional complex species}

\item{exa}{A vector of the ion size parameters for the complexes of the additional complex species}

\item{exK}{A vector of the log K of the dissociation constants of the additional complex species}

\item{bal}{Species to balance against (Defaults to Cl, can also be set to NULL for none)}
}
\value{
A list containing the concentrations, activity coefficients, and pH at equilibrium
}
\description{
Mass balance and charge balance solver for chemical equilibria at 0.5 kb. This is a wrapper function for chemsolve_generic and prods.
}
\details{
A wrapper for the chemsolve_generic function that allow easy addition of product species.
Use the generic function (chemsolve_generic) if new basis species need to be added or if the log K/temperature range is extended (up or down). \cr
Normally total moles anions = total moles cations. The charge balance (without any speciation) is adjusted to zero 
by balancing against Cl- (default) as otherwise the calculation is too sensitive to H+. The balancing species can be easily changed to any of the other basis species (or none). \cr
Choose reasonable starting values; for H+, OH- and equilibrium concentrations of the basis species. \cr
If negative concentrations are calculated,  choose better initial starting values. \cr
To exclude a  basis value set the basis concentration to zero and the concentrations of this and the derived species 
will be vanishing small and can be ignored. In the generic version of the program the basis species are simply left out. \cr
The basis species are:  Na+, K+, Mg2+, Ca2+, Cl-, SO42-. The default complexes are:  NaCl°, KCl°, HCl°, KOH°, NaOH°, KSO4-, NaSO4-,HSO4-,CaSO4°,MgSO4°, MgCl+,CaCl+,CaCl2°,MgOH+,CaOH+. 
Additional complexes based on the existing basis species can easily be added (see example below). \cr
To set up a problem in chemsolve or chemsolve_generic it is relatively trivial to set up and manipulate a sequence of commands in Excel and cut and past into R.
Complex dissociation constants (log K)  are from SupCrt 92  slop98.dat \url{http://geopig.asu.edu/?q=tools} \cr
The Debye_Hückel parameters (A, B) equations are polynomial fits to data at 0.5 kb from tables in Helgeson & Kirkham (1974).\cr
Note Bdot is not used.\cr
Helgeson H. C. and Kirkham D. H. (1974) Theoretical prediction of the thermodynamic behavior of aqueous electrolytes at high pressures and temperatures: II. Debye-Hückel parameters for activity coefficients and relative partial molar properties American Journal of Science 274, 1199-1261. \cr
}
\examples{
## Add KHSO4° as an additional complex given the existing list of basis species and 
## calculate the equilibrium concentrations and pH at 400°C.
chemsolve(exprod =  c("KHSO4"), exconstit = c("K", "H", "SO4"),
exnumz =  c(3), excharges =  c(0), exa =  c(0),  exK =  c(-8.701), bal = "Cl")
## Determine the equilibria at a range of temperatures. 
## Additional complexes can be added per the previous example.
temps <- seq(300,400,10) #A vector of temperatures repeating every 10 degrees from 300 to 400
## Na concentration is changed slightly from the default. Further parameters can be added.
r <- lapply(temps,chemsolve,Nat=0.45) #Creates a list of the results
r[[1]] #Display results from first temperature
r[[10]] #Display the results of the 10th temperature
}
